refactor(ai-sidecar): Hoist ws_commands into a shared package
The WebSocket<->ACP stdio bridge is not specific to any model vendor: every
sidecar in this tree needs the same framing to talk to the Jaeger AI gateway.
Moving it to scripts/ai-sidecar/shared/ makes "all sidecars present an identical
ACP surface" structural instead of something two copies of a file happen to
agree on, ahead of a second sidecar landing alongside the Gemini one.

The move is mechanical — ws_commands.py is byte-for-byte unchanged. Because
these sidecars run as directories of scripts rather than installed packages,
main.py (normal runs) and a new conftest.py (pytest) put scripts/ai-sidecar on
sys.path, and pyright's extraPaths points at the same directory.

Verified with the Gemini sidecar's own CI gates: `uv run pyright ./` is clean
and all 9 tests in `uv run pytest -q` pass.
