+++ scripts/ai-sidecar/shared/ws_commands.py
